Nested signature verification
